fix a LOT of issues, but not all of them

+++ src/V1/Parsing/Common/OrientationField.php
@@ -20,7 +20,7 @@ class OrientationField extends BaseField
     public $value;
 
     /**
-     * @param array $rawPrediction Raw prediction array.
+     * @param array<string, mixed> $rawPrediction Raw prediction array.
      * @param integer|null $pageId Page number for multi pages document.
      * @param boolean $reconstructed Whether the field was reconstructed.
      * @param string $valueKey Key to use for the value.
@@ -34,7 +34,7 @@ class OrientationField extends BaseField
         parent::__construct($rawPrediction, $pageId, $reconstructed, $valueKey);
         $this->value = 0;
         if (array_key_exists($valueKey, $rawPrediction) && is_numeric($rawPrediction[$valueKey])) {
-            $this->value = (float) ($rawPrediction[$valueKey]);
+            $this->value = (int) ($rawPrediction[$valueKey]);
             if (!in_array($this->value, [0, 90, 180, 270], true)) {
                 $this->value = 0;
             }
